# backend/src/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PeriodName": {
				"enum": [
					"Period 0",
					"Period 1",
					"Period 2"
				],
				"type": "string"
			},
			"Level": {
				"enum": [
					"A1X",
					"G1X",
					"G2X"
				],
				"type": "string"
			},
			"Ecv": {
				"enum": [
					"C",
					"C/E",
					"E",
					"V"
				],
				"type": "string"
			},
...
			"PlannedCourseSpecialization": {
				"properties": {
					"name": {
						"type": "string"
					},
					"ECV": {
						"$ref": "#/components/schemas/Ecv"
					}
				},
				"required": [
					"name",
					"ECV"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlannedCourse": {
				"properties": {
					"courseCode": {
						"type": "string"
					},
					"courseName": {
						"type": "string"
					},
					"credits": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/Level"
					},
					"timetableModule": {
						"type": "string"
					},
					"ECV": {
						"$ref": "#/components/schemas/Ecv"
					},
					"info": {
						"type": "string"
					},
					"semester": {
						"type": "string",
						"nullable": true
					},
					"specializations": {
						"items": {
							"$ref": "#/components/schemas/PlannedCourseSpecialization"
						},
						"type": "array"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"courseCode",
					"courseName",
					"credits",
					"level",
					"timetableModule",
					"ECV",
					"info",
					"semester",
					"specializations",
					"note"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PeriodPlan": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/PeriodName"
					},
					"courses": {
						"items": {
							"$ref": "#/components/schemas/PlannedCourse"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"courses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SemesterPlan": {
				"properties": {
					"name": {
						"type": "string"
					},
					"periods": {
						"items": {
							"$ref": "#/components/schemas/PeriodPlan"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"periods"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MasterPlan": {
				"properties": {
					"programName": {
						"type": "string"
					},
					"startYear": {
						"type": "string"
					},
					"semesters": {
						"items": {
							"$ref": "#/components/schemas/SemesterPlan"
						},
						"type": "array"
					},
					"specializion": {
						"type": "string"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"programName",
					"startYear",
					"semesters",
					"specializion",
					"note"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MasterPlanResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"plan": {
						"$ref": "#/components/schemas/MasterPlan"
					},
					"user": {
						"$ref": "#/components/schemas/UserDTO"
					}
				},
				"required": [
					"id",
					"title",
					"plan",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMasterPlanRequestDTO": {
				"properties": {
					"title": {
						"type": "string",
						"minLength": 8
					},
					"plan": {
						"$ref": "#/components/schemas/MasterPlan"
					}
				},
				"required": [
					"title",
					"plan"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMasterPlanRequestDTO": {
				"properties": {
					"title": {
						"type": "string",
						"minLength": 8
					},
					"plan": {
						"$ref": "#/components/schemas/MasterPlan"
					}
				},
				"required": [
					"title",
					"plan"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Api-Key",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "big-backend",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
...
		"/plan": {
			"get": {
				"operationId": "GetAllMasterPlans",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MasterPlanResponseDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMasterPlan",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMasterPlanRequestDTO"
							}
						}
					}
				}
			}
		},
		"/plan/user": {
			"get": {
				"operationId": "GetAllMasterPlanForUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MasterPlanResponseDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/plan/{id}": {
			"put": {
				"operationId": "UpdateMasterPlan",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMasterPlanRequestDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}